Nicolas Regnault: - correct a problem in NDMAPSpinChain that produced a compilation error
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@50 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ src/HilbertSpace/QHEHilbertSpace/FermionOnSphere.h
@@ -60,7 +60,7 @@ class FermionOnSphere :  public ParticleOnSphere
   // maximum shift used for searching a position in the look-up table
   int MaximumLookUpShift;
   // memory used for the look-up table in a given lzmax sector
-  int LookUpTableMemorySize;
+  unsigned long LookUpTableMemorySize;
   // shift used in each lzmax sector
   int* LookUpTableShift;
   // look-up table with two entries : the first one used lzmax value of the state an the second
